+++ packages/core/right-side-panel/src/components/RightSidePanel/index.tsx
@@ -28,6 +28,16 @@ export interface Props {
   attachPanelTo: string;
   // Right Hand Side panel content
   children?: ReactNode;
+  // Don't animate the component when the component is mounted (false by default)
+  skipAnimationOnMount?: boolean;
+  // Mount component on enter (true by default)
+  mountOnEnter?: boolean;
+  // Unmount component on exit (true by default)
+  unmountOnExit?: boolean;
+  // Disable enter animation (false by default)
+  disableEnterAnimation?: boolean;
+  // Disable exit animation (false by default)
+  disableExitAnimation?: boolean;
   // Function to be executed when the open animation finishes
   onOpenAnimationFinished?: () => void;
   // Function to be executed when the close animation finishes
@@ -71,6 +81,11 @@ export class RightSidePanel extends Component<Props, State> {
     const {
       children,
       isOpen,
+      skipAnimationOnMount = false,
+      mountOnEnter = true,
+      unmountOnExit = true,
+      disableEnterAnimation = false,
+      disableExitAnimation = false,
       onOpenAnimationFinished,
       onCloseAnimationFinished,
     } = this.props;
@@ -79,9 +94,11 @@ export class RightSidePanel extends Component<Props, State> {
       <Transition
         in={isOpen}
         timeout={transitionDurationMs}
-        mountOnEnter
-        unmountOnExit
-        appear
+        mountOnEnter={mountOnEnter}
+        unmountOnExit={unmountOnExit}
+        appear={!skipAnimationOnMount}
+        enter={!disableEnterAnimation}
+        exit={!disableExitAnimation}
         onEntered={onOpenAnimationFinished}
         onExited={onCloseAnimationFinished}
       >