+++ typings/atlaskit.d.ts
@@ -14,11 +14,8 @@ declare module '@atlaskit/form';
 declare module '@atlaskit/global-navigation';
 // TODO - Add icon import paths to entry-points.tsconfig.json
 declare module '@atlaskit/icon/*';
-declare module '@atlaskit/icon-file-type';
 declare module '@atlaskit/icon-file-type/*';
-declare module '@atlaskit/icon-object';
 declare module '@atlaskit/icon-object/*';
-declare module '@atlaskit/icon-priority';
 declare module '@atlaskit/icon-priority/*';
 declare module '@atlaskit/input';
 declare module '@atlaskit/item';
@@ -43,6 +40,7 @@ declare module '@atlaskit/textfield';
 declare module '@atlaskit/theme/colors';
 
 // Build
+declare module '@atlaskit/icon-build-process';
 declare module '@atlaskit/ssr';
 declare module '@atlaskit/docs';
 declare module '@atlaskit/build-utils';