+++ packages/core/comment/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @atlaskit/comment
 
+## 9.0.1
+
+### Patch Changes
+
+- [patch][35d2229b2a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/35d2229b2a):
+
+  Adding missing license to packages and update to Copyright 2019 Atlassian Pty Ltd.
+
+## 9.0.0
+
+### Major Changes
+
+- [major][2376ada634](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/2376ada634):
+
+  Comment has been converted to Typescript. Typescript consumers will now get static type safety. Flow types are no longer provided. No API or behavioural changes.
+
+## 8.0.15
+
+### Patch Changes
+
+- [patch][a2d0043716](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a2d0043716):
+
+  Updated version of analytics-next to fix potential incompatibilities with TS 3.6
+
 ## 8.0.14
 
 - Updated dependencies [166eb02474](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/166eb02474):