+++ packages/core/toggle/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @atlaskit/toggle
 
+## 8.1.2
+
+### Patch Changes
+
+- [patch][35d2229b2a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/35d2229b2a):
+
+  Adding missing license to packages and update to Copyright 2019 Atlassian Pty Ltd.
+
+## 8.1.1
+
+### Patch Changes
+
+- [patch][a2d0043716](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a2d0043716):
+
+  Updated version of analytics-next to fix potential incompatibilities with TS 3.6
+
+## 8.1.0
+
+### Minor Changes
+
+- [minor][decb366d7a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/decb366d7a):
+
+  Adding an optional prop `testId` that will set the attribute value `data-testid`. It will help products to write better integration and end to end tests.
+
 ## 8.0.5
 
 ### Patch Changes