+++ packages/elements/analytics-gas-types/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @atlaskit/analytics-gas-types
 
+## 4.0.12
+
+### Patch Changes
+
+- [patch][35d2229b2a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/35d2229b2a):
+
+  Adding missing license to packages and update to Copyright 2019 Atlassian Pty Ltd.
+
+## 4.0.11
+
+### Patch Changes
+
+- [patch][a2d0043716](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a2d0043716):
+
+  Updated version of analytics-next to fix potential incompatibilities with TS 3.6
+
 ## 4.0.10
 
 ### Patch Changes